+++ packages/gloves-client/src/index.ts
@@ -308,7 +308,7 @@ class McpSession {
         },
         clientInfo: {
           name: "@gloves/client",
-          version: "0.1.0",
+          version: "0.1.1",
         },
         _meta: {
           sessionToken: token,
@@ -366,7 +366,7 @@ class McpSession {
         },
         clientInfo: {
           name: "@gloves/client",
-          version: "0.1.0",
+          version: "0.1.1",
         },
         _meta: {
           sessionToken: token,